fix: remove `NamingServiceAwareTraitTest` due to over-coupling
The trait depends on `$this->getRepo()` which doesn't exist in this trait so would require this trait to be used in combination with classes that provide this method.
