+++ phpstan.neon.dist
@@ -1,7 +1,4 @@
 parameters:
-  level: 1
+  level: 0
   paths:
     - module
-    - public
-    - config
-    - test
